Problems with alternate entries and returns (#347)
* Problem with alternate entries

Compilation of this program with an alternate entry fails:

    call ee(17)
  end

  subroutine ss(nn)
    print*, nn
    return
  entry ee(nn)
    print*, nn
  end

Compilation succeeds if the program is moved below the subroutine.

The problem is that with the code as listed, the front end provides
ProcEntityDetails for the definition of entry ee (from the call before the
entry is visited), when it should provide more complete SubprogramDetails.
The front end is generating the SubprogramDetails, but it doesn't retain
them.  The fix in resolve-names.cpp is to always retain the SubprogramDetails.

With this fix, the "2 wrongs make a right" workaround for the function entry
point instance of this problem in PFTBuilder.cpp is no longer needed.