enable fir.box in StoreOp and EmboxOp usage in GlobalOp body
- Enable and implement codegen for fir.store of fir.box
This is done by loading the value before storing it.

- Enable fir.embox usage in global initializer

In executable code, fir.embox is rewritten to llvm as a temp allocation
for the descriptor followed by stores to fill up the fields.
This cannot be done in global initialization code because alloca and stores
are illegal there.
This patch adds a way to lower embox to a undefOp + insertValueOp instead of
the allocaOp + storeOp. This is done by providing a call back to embox codegen
that writes descriptor fields.
When the context is inside a globalOp, then the callback use insertValueOp,
otherwise, the callback use storeOp.

Additionally, the CodeGenRewrite pre-pass had to be modified so that it also runs
on global ops initializer. This pre-pass is responsible for getting rid of the
fir.shapeOp and rewriting array embox to Xembox. The pass was previously only
run on functions bodies.

- fix BoxEleSizeOp codegen (bad type was passed to the GEP)