[NFC] Moves command line options that control passes.
This code shuffling is for upstreaming. Unlike clang and llvm, command-line
options are not placed next to the pass itself. Since we want disable switches
to turn off passes for debugging, it doesn't make sense to convert these to
"pass options" (per MLIR) as the goal is to disable and not run the pass.