Enable transformational intrinsic function lowering
Use the normal (scalar) intrinsic all path to deal with transformational
intrinsic that returns arrays. The intrinsic framework already handles
returning arrays. The runtime is allocating the result storage.

- At a high level, detect transformational to avoid going through
asArray and avoid creating an additional temp + copy. This path is
used when the transformational does not appear inside an elemental
expression/assignment (for instance, when it the result is a function
argument).

- In array genProcRef, forward transformational to the scalar framework
to create the result extended value, and use genarr on this extended
value results. This path is used when the transformational result
appear in elemental expressions/assignment.

Tests are based on ALL, and are disabled until ALL lowering
is merged.