Fixes bug #783.
Instead of using a nullptr to signal that an array's column is assumed
size, use a fir.undefined value.  The undefined value is semantically
what is meant and it is an actual ssa-value in the IR, so it can be used
without crashing the compiler with sigsegv failures. Operationally, it is
incorrect to use the "default" column size of an assumed-size array, so
the use of fir.undefined will make this explicit in the FIR.