Fix contiguous local pointer assignment from fir.box
Scalar and contiguous pointer assignment may be lowered to
a set of local variable instead of a fir.box, in which case they
should not be updated with a fir.box. It is possible that the
RHS of assignments to such pointers is lowered to a fir.box
because the contiguous aspects of the RHS may not be obvious or
guaranteed in lowering.
Add handling for this case by reading the fir.box to implement the
pointer assignment.

On the way, fix two related issues:
 - missing casts before store when updating the local variables
   representing the pointer. There were i64 to index store issues
   otherwise in some cases.
 - Alloca/Allocmem did not allow to allocate a fir.ptr<fir.array<?x...xT>>
   which is needed to track the address of a contiguous local array
   pointer. The alloc verify was going through the fir.ptr and then
   complained that not extents was provided in the alloc for the
   fir.array type. This looks bogus to me, and I do not see the
   purpose of this check, so I removed it.