Fix passing derived pointers/allocatables as implicit argument
So far genExtAddr on pointers/allocatables is conservative and does
not get rid of the fir.box when lowering represented via a fir.box
because with the current lowering/FIR design, it has no way to know
if a fir::BoxValue is polymorphic or not. If it is polymorphic,
the fir.box will be required to get the dynamic type in later uses.

This caused issue around copy-in/copy-out where the code did not
expect to find a box after lowering of an expression previously
determined as simply contiguous.

Simply read the address from the fir.box after genExtAddr in those
cases.