[flang] Fix the SHIFT argument of the CSHIFT intrinsic
While working on EOSHIFT, I realized that the lowering code for the CSHIFT
intrinsic was incorrect.  The SHIFT argument was being passed as an address
but was being treated as an integer value.  My tests were accidentally working
since the runtime code was taking the modulus of the address and coincidentally
getting the correct value.