Make extern weak symbols for type descriptor not defined in translation unit
If a derived type is defined in a module in another translation unit,
its "<scope mangling>.dt.<typename>" global is not declared in the
current translation unit in lowering, and therefore not accessible in
codegen. The symbol is not lowered because it is not use associated in
any of the program units lowered in the current translation unit. If we wanted
to deal with this in lowering, we would need to start lowering
declaration for all variables from the module used by the lowered
program units, which is not the current approach in lowering).

Declare a tentative (extern weak) type descriptor symbol in codegen
if no global exists for the type descriptor symbol. It may be resolved
at link time, and will otherwise be a nullptr (so far, do not make it a
link error if the symbol is defined nowhere to support codegen of FIR
files produced without runtime info symbols for debug purposes).