FORALL lowering
This is a major restructuring of how FORALL loops are lowered in the bridge from
parse trees to FIR. Previously, FORALL loops had been lowered like any other
structured control-flow loops and then as a more dynamic form of array
expression.

This refactoring is a middle road approach where the FORALL context is analyzed
and then lowered in the bridge. This allows lowering to create loop nests that
correctly deal with impure expressions and pure masking expressions, which were
not aligned with array expressions.  Lowering of assignment statements in the
FORALL context is then done within the pre-built context by expression lowering.
Secondly, expression lowering is refactored into a new "scalar array" traversal
and code gen that works with array values as instantiated by the bridge in the
earlier loop generation phase.

Switching to this structure allows the array expression framework to be
simplified since it no longer is required to deal with "special" scalar
expressions and try to raise them into an array form.

Allocatable arrays and pointer assignments are supported.

User-defined ASSIGNMENT in (scalar) FORALL contexts is supported.

This also introduces several correctness fixes.

This does not correct known problems with arrays of CHARACTER, incomplete
information with respect to derived types, elemental user-defined ASSIGNMENT in
array expressions, PDTs, etc.