[flang] Fix build failures caused by bogus GNU C++ warnings
When I do out-of-tree builds in fir-dev, I get warnings when building
lib/Parser/parsing.cpp that something may be uninitialized.  I don't get
these warnings on in-tree builds or on any builds from llvm-project.

I Googled about this and it looks like it's a bug in the GNU C++
compiler.  Adding the "-Wno-maybe-unitialized" option fixes this and
doesn't break anything else.