[flang]Pass to add vscale range attribute
Add vscale range attirbute for the Scalable Vector Extension (SVE)
if provided on the command-line (options in a previous commit)

A pass is used to add the atribute to all functions. The
vectorizer will use this attribute to generate the SVE instruction
to match the range specified. The attribute is harmless if there
is no vectorizable operations in the function.